+++ packages/xy/tests/scatter/ScatterCrosshair.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import { Chart, View } from '@chsk/core'
-import { Scatter, ScatterCrosshair, ScatterPoints } from '../../src'
+import { Scatter, ScatterCrosshair, ScatterInteractiveDataItem, ScatterPoints } from '../../src'
 import { scatterProps } from './scatter.props'
 import { getNumberAttr } from '../../../core/tests/utils'
 
@@ -41,7 +41,7 @@ describe('ScatterCrosshair', () => {
         expect(screen.queryByRole('scatter-crosshair')).toBeNull()
     })
 
-    it('draws two lines as crosshairs', () => {
+    it('draws two crosshair lines', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
@@ -59,7 +59,7 @@ describe('ScatterCrosshair', () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
-                    <ScatterCrosshair variant={'vertical'} />
+                    <ScatterCrosshair visible={[true, false]} />
                 </Scatter>
             </Chart>
         )
@@ -67,13 +67,29 @@ describe('ScatterCrosshair', () => {
         fireEvent.mouseMove(detector, { clientX: 50, clientY: 50 })
         const lines = screen.getByRole('scatter-crosshair').querySelectorAll('line')
         expect(lines).toHaveLength(1)
+        expect(lines[0].getAttribute('x1')).toEqual(lines[0].getAttribute('x2'))
+    })
+
+    it('draws one line as horizontal crosshair', () => {
+        render(
+            <Chart>
+                <Scatter {...scatterProps}>
+                    <ScatterCrosshair visible={[false, true]} />
+                </Scatter>
+            </Chart>
+        )
+        const detector = screen.getByRole('crosshair-detector')
+        fireEvent.mouseMove(detector, { clientX: 50, clientY: 50 })
+        const lines = screen.getByRole('scatter-crosshair').querySelectorAll('line')
+        expect(lines).toHaveLength(1)
+        expect(lines[0].getAttribute('y1')).toEqual(lines[0].getAttribute('y2'))
     })
 
     it('removes lines upon mouse leave', async () => {
         render(
             <Chart>
                 <Scatter {...scatterProps}>
-                    <ScatterCrosshair variant={'horizontal'} />
+                    <ScatterCrosshair variant={'y'} />
                 </Scatter>
             </Chart>
         )
@@ -82,8 +98,7 @@ describe('ScatterCrosshair', () => {
         await waitFor(() => {
             expect(screen.queryByRole('crosshair-presence')).toBeDefined()
         })
-        const lines = screen.getByRole('scatter-crosshair').querySelectorAll('line')
-        expect(lines).toHaveLength(1)
+        expect(screen.getByRole('scatter-crosshair').querySelectorAll('line')).toHaveLength(2)
         fireEvent.mouseLeave(detector)
         await waitFor(() => {
             expect(screen.queryByRole('crosshair-presence')).toBeNull()
@@ -104,8 +119,7 @@ describe('ScatterCrosshair', () => {
         await waitFor(() => {
             expect(screen.queryByRole('crosshair-presence')).toBeDefined()
         })
-        const lines = screen.getByRole('scatter-crosshair').querySelectorAll('line')
-        expect(lines).toHaveLength(2)
+        expect(screen.getByRole('scatter-crosshair').querySelectorAll('line')).toHaveLength(2)
         // fire an event very far from the data points, crosshair should disappear
         fireEvent.mouseMove(detector, { clientX: 25000, clientY: 25000 })
         await waitFor(() => {
@@ -121,8 +135,7 @@ describe('ScatterCrosshair', () => {
                 data={{ disabledKeys: new Set<string>(['quadratic']) }}
             >
                 <Scatter {...scatterProps}>
-                    <ScatterPoints />
-                    <ScatterCrosshair variant={'horizontal'} />
+                    <ScatterCrosshair variant={'y'} visible={[false, true]} />
                 </Scatter>
             </Chart>
         )
@@ -137,5 +150,30 @@ describe('ScatterCrosshair', () => {
         const line = screen.getByRole('scatter-crosshair').querySelector('line')
         expect(getNumberAttr(line, 'y1')).toBeGreaterThan(150)
         expect(getNumberAttr(line, 'y2')).toBeGreaterThan(150)
+    })
+
+    it('provides point data to onClick handler', async () => {
+        let data: ScatterInteractiveDataItem | undefined = undefined
+        const handleClick = (x?: ScatterInteractiveDataItem) => {
+            data = x
+        }
+        render(
+            <Chart size={[400, 300]} padding={[0, 0, 0, 0]}>
+                <Scatter {...scatterProps}>
+                    <ScatterCrosshair handlers={{ onClick: handleClick }} />
+                </Scatter>
+            </Chart>
+        )
+        const detector = screen.getByRole('crosshair-detector')
+        // trigger the detector near the top of the chart, i.e. near quadratic series
+        fireEvent.mouseMove(detector, { clientX: 200, clientY: 15 })
+        await waitFor(() => {
+            expect(screen.queryByRole('crosshair-presence')).toBeDefined()
+        })
+        fireEvent.click(detector, { clientX: 200, clientY: 15 })
+        await waitFor(() => {
+            expect(data).not.toBeUndefined()
+            expect(data?.id).toBe('quadratic')
+        })
     })
 })